# Vectorize CosineSimilarityImplementation with Vector<T>